Updated projects to target net8
