+++ .github/workflows/ci-main.yml
@@ -7,8 +7,8 @@ on:
     types: [ published ]
 
 env:
-    NETCORE_VERSION: '7.0.x'
-    OUTPUT_DIR: 'net7.0'
+    NETCORE_VERSION: '8.0.x'
+    OUTPUT_DIR: 'net8.0'
     DOTNET_SKIP_FIRST_TIME_EXPERIENCE: true
     DOTNET_CLI_TELEMETRY_OPTOUT: true
     PROJECT_NAME: LocalStorage
