+++ .github/workflows/ci-pr.yml
@@ -5,7 +5,7 @@ on:
     branches: [ main ]
 
 env:
-    NETCORE_VERSION: '7.0.100'
+    NETCORE_VERSION: '8.0.100'
     PROJECT_NAME: Blazored.LocalStorage
 
 jobs:
