+++ Directory.Build.props
@@ -4,6 +4,7 @@
   <PropertyGroup Label="Package Versions">
     <DotNet6Version>6.0.0</DotNet6Version>
     <DotNet7Version>7.0.0</DotNet7Version>
+    <DotNet8Version>8.0.0</DotNet8Version>
   </PropertyGroup>
 
 </Project>
