--- samples/BlazorServer/Pages/_Host.cshtml
@@ -1,9 +1,0 @@
-﻿@page "/"
-@namespace BlazorServer.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@{
-    Layout = "_Layout";
-}
-
-<component type="typeof(App)" render-mode="ServerPrerendered" />
-
