--- samples/BlazorServer/_Imports.razor
@@ -1,12 +1,0 @@
-@using System.Net.Http
-@using Microsoft.AspNetCore.Authorization
-@using Microsoft.AspNetCore.Components.Authorization
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.JSInterop
-@using BlazorServer 
-@using BlazorServer.Shared
-
-@using Blazored.LocalStorage
