+++ samples/BlazorWebAssembly/Shared/NavMenu.razor
@@ -19,12 +19,11 @@
 </div>
 
 @code {
-    private bool collapseNavMenu = true;
+    private bool _collapseNavMenu = true;
 
-    private string NavMenuCssClass => collapseNavMenu ? "collapse" : null;
+    private string? NavMenuCssClass 
+        => _collapseNavMenu ? "collapse" : null;
 
-    private void ToggleNavMenu()
-    {
-        collapseNavMenu = !collapseNavMenu;
-    }
+    private void ToggleNavMenu() 
+        => _collapseNavMenu = !_collapseNavMenu;
 }
