+++ samples/InteractiveServer/NewtonSoftJsonSerializer.cs
@@ -1,0 +1,13 @@
+using Blazored.LocalStorage.Serialization;
+using Newtonsoft.Json;
+
+namespace BlazorServer;
+
+public class NewtonSoftJsonSerializer : IJsonSerializer
+{
+    public T? Deserialize<T>(string text)
+        => JsonConvert.DeserializeObject<T>(text);
+
+    public string Serialize<T>(T obj)
+        => JsonConvert.SerializeObject(obj);
+}
